requirements.txt
flask
flask-cors
chromastone
reportlab
pymongo
python-dotenv
